# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Furtalk 评论系统的 HTTP API：覆盖站点与管理员引导、\n第一方与 widget 认证/会话、评论发布与审核、站点/用户/设置/\n提供商管理与邮件通知退订。",
        "title": "Furtalk API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/bootstrap/admin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "bootstrap"
                ],
                "summary": "创建首位管理员",
                "parameters": [
                    {
                        "description": "初始化令牌与管理账号",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BootstrapAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "初始化完成"
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "邮箱已被注册",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "初始化已不可用",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/captcha/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "获取按 action 划分的公共 CAPTCHA 配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CAPTCHA 业务 action（如 password_login、comment）",
                        "name": "action",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "公共 CAPTCHA 配置（不含机密）",
                        "schema": {
                            "$ref": "#/definitions/handler.CaptchaConfigResponse"
                        }
                    },
                    "422": {
                        "description": "action 缺失或超长"
                    },
                    "503": {
                        "description": "验证码服务暂不可用"
                    }
                }
            }
        },
...
        "/api/v1/notification-unsubscriptions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "退订通知邮件",
                "parameters": [
                    {
                        "description": "邮件中的签名退订 token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UnsubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "退订完成"
                    },
                    "400": {
                        "description": "退订令牌无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.BootstrapAdminRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "setup_token": {
                    "type": "string"
                }
            }
        },
...
        "handler.CaptchaConfigResponse": {
            "type": "object",
            "properties": {
                "captcha": {
                    "$ref": "#/definitions/handler.PublicCaptchaConfig"
                },
                "required": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.PublicCaptchaConfig": {
            "type": "object",
            "properties": {
                "api_endpoint": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "site_key": {
                    "type": "string"
                }
            }
        },
...
        "handler.UnsubscribeRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorBody"
                }
            }
        },
...
    }
}